+++ test/integration/tests/796-ghc-options/Main.hs
@@ -1,0 +1,7 @@
+import StackTest
+
+main :: IO ()
+main = do
+    stack ["build"]
+    stackErr ["build", "--ghc-options=-DBAZ"]
+    stack ["build", "--ghc-options=-DQUX"]