+++ ChangeLog.md
@@ -4,6 +4,10 @@ Major changes:
 
 * You now have more control over how GHC versions are matched, e.g. "use exactly this version," "use the specified minor version, but allow patches," or "use the given minor version or any later minor in the given major release." The default has switched from allowing newer later minor versions to a specific minor version allowing patches. For more information, see [#736](https://github.com/commercialhaskell/stack/issues/736) and [#784](https://github.com/commercialhaskell/stack/pull/784).
 
+Other enhancements:
+
+* Added the `ghc-options` field to stack.yaml [#796](https://github.com/commercialhaskell/stack/issues/796)
+
 Bug fixes:
 
 * Ignore disabled executables [#763](https://github.com/commercialhaskell/stack/issues/763)